backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
Pillow==10.3.0
pydantic==2.7.4
pydantic-settings==2.3.4
